# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API documentation for my application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/saveuserpreferences": {
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Save user news preferences",
                "description": "Stores the user's preferred news category, source, and author.",
                "operationId": "a9725241b8174b6176e7bb0821ebfb07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "business",
                                        "nullable": true
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "CNN",
                                        "nullable": true
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "John Doe",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User preferences saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User preferences saved successfully!"
                                        },
                                        "preference": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "business"
                                                },
                                                "source": {
                                                    "type": "string",
                                                    "example": "CNN"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-02-18T19:16:24.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-02-18T19:16:24.000000Z"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "category": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The category field must be a string."
                                                    ]
                                                },
                                                "source": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The source field must be a string."
                                                    ]
                                                },
                                                "author": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The author field must be a string."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too many requests, please slow down."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/getuserpreferencesnews": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preference-based news",
                "description": "Fetches news articles based on the user's saved preferences (category, source, author). If no preferences are found, returns all news articles.",
                "operationId": "a4d918695dfd58d722084b8787d61470",
                "responses": {
                    "200": {
                        "description": "Preferred news articles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferred news articles retrieved successfully!"
                                        },
                                        "articles": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "State Department had a plan to buy $400M worth of armored Tesla vehicles"
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "mlive.com"
                                                            },
                                                            "source": {
                                                                "type": "string",
                                                                "example": "Biztoc.com"
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "example": "business"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "State Department was in talks with Tesla for armored EVs."
                                                            },
                                                            "url": {
                                                                "type": "string",
                                                                "format": "uri",
                                                                "example": "https://example.com/article"
                                                            },
                                                            "published_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-02-17 18:37:00"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-02-18T17:13:08.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-02-18T17:13:08.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8082/api/getuserpreferencesnews?page=1"
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8082/api/getuserpreferencesnews?page=2",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too many requests, please slow down."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}